requirements.txt
# Web scraping
requests
beautifulsoup4
selenium==4.25.0
webdriver-manager

# Data handling
pandas

# Google Cloud BigQuery
google-cloud-bigquery
pandas-gbq
google-auth